Major cosmetic changes.
Added marquee for problem name when name is greater than enclosing
column width. Did not add marquee for the timer string yet. Timer string
keeps duplicating for every problem on a new line, maybe because of the
use of setInterval function.
Added a max-height to container and a scroll-bar shows if the max-height
is reached.

Next: Create custom scroll-bar, thinner and better looking than the
default one.